A new class FileManager has been implemented. The code related to creating and managing files has been moved there.
