--- src/jetbrains/Command.java
@@ -1,4 +1,0 @@
-package jetbrains;
-
-public class Command {
-}
